% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calibrate.unconditional}
\alias{calibrate.unconditional}
\title{Unconditional calibration}
\usage{
calibrate.unconditional(
  m,
  n,
  distribution_key,
  distribution_parameters,
  monitoring_statistic_type,
  monitoring_statistic_parameters,
  limits_seq,
  is_upper_limit,
  nsim,
  run_length_cap
)
}
\arguments{
\item{m}{The dimension used for the reference sample}

\item{n}{The dimension used for the test samples}

\item{distribution_key}{A string which identifies a distribution or a distribution family. Used in combination
with the "distribution_parameters" parameter in order to select the IC distribution. See details for consulting
the available distributions and the respective keys.}

\item{distribution_parameters}{A list with the distribution parameters (an empty list should be supplied if the distribution has not parameters)}

\item{monitoring_statistic_type}{Either "npc" or "simple_statistic"}

\item{monitoring_statistic_parameters}{A list with type-specific parameters of the monitoring statistic. For the
"simple_statistic" type the only parameter is "statistic" which should be set to a valid statistic key.
For the "npc" type the four (required) parameters are "statistics", "permutation_distributions", "tails" and
"combining_function". The "statistics", "permutation_distributions" and "tails" parameters should be three lists of the
same length. The "combining_function" parameters should be a string. See details for consulting the available statistics and combining
functions and the keys to be used.}

\item{limits_seq}{The numeric vector of limits for which the run length should be recorded at each simulation}

\item{is_upper_limit}{A boolean parameter used to select whether the chart uses an upper limit or a lower limit}

\item{nsim}{The number of simulations}

\item{run_length_cap}{A limit for the run length in the simulations used to guarantee convergence of the algorithm}
}
\value{
A numeric matrix of size nsim x length(limits_seq)
}
\description{
Unconditional calibration for the Stehwart-type charts implemented by this package
}
\details{
The supported distributions and the respective parameters are:
\itemize{
 \item{Normal - key:"norm", params:"mean" and "sd"}
 \item{Laplace - key:"laplace", params:"location" and "scale"}
 \item{Student's T - key:"t", params:"df"}
 \item{Cauchy - key:"laplace", params:"location" and "scale"}
 \item{Chi squared - key:"chi_squared", params:"df"}
 \item{Exp(lambda=1) shifted by -1 - key:"normalized_rate_one_exponential"}
 \item{Gamma(scale=4,shape=2) shifted by -2 - key:"centered_gamma_with_shape_equal_four"}
}
The supported statistics and the respective keys to be used are:
\itemize{
 \item{Wilcoxon rank sum - "wilcoxon_rank_sum"}
 \item{Mann-Whitney - "mann_whitney"}
 \item{Mean-normalized wilcoxon rank sum  - "centered_wilcoxon_rank_sum"}
 \item{ Percentile modified rank test with s=r=0.42 - "percentiles_42_42"}
 \item{ Percentile modified rank test with s=r=0.46 - "percentiles_46_46"}
 \item{Klotz - "klotz_statistic"}
 \item{Ansari Bradely - "ab_statistic"}
 \item{Lepage - "lepage"}
 \item{Cucconi - "cucconi"}
}
The three options for the values of the "tails" list parameter of npc monitoring statistics are: "two_sided", "left" and "right".
The three options for the "combining_function" parameter of npc monitoring statistics are: "fisher", "liptak" and "tippet"
}
