% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evaluate.unconditional}
\alias{evaluate.unconditional}
\title{Unconditional evaluation}
\usage{
evaluate.unconditional(
  m,
  n,
  limits,
  location_shift,
  scale_multiplier,
  distribution_key,
  distribution_parameters,
  monitoring_statistic_type,
  monitoring_statistic_parameters,
  nsim,
  run_length_cap
)
}
\arguments{
\item{m}{The dimension used for the reference sample}

\item{n}{The dimension used for the test samples}

\item{distribution_key}{A string which identifies a distribution or a distribution family. Used in combination with the "distribution_parameters" parameter in order to select the IC distribution}

\item{distribution_parameters}{A list with the distribution parameters (an empty list should be supplied if the distribution has not parameters)}

\item{monitoring_statistic_type}{Either "npc" or "simple_statistic"}

\item{monitoring_statistic_parameters}{Type-specific parameters of the monitoring statistic. For the "simple_statistic" type
the only parameter is "statistic" which should be set to a valid statistic key. For the "npc" type the three
(required) parameters are "statistics", "permutation_distributions" and "tails"}

\item{nsim}{The number of simulations}

\item{run_length_cap}{A limit for the run length in the simulations used to guarantee convergence of the algorithm}
}
\description{
Unconditional evaluation for Stehwart-type charts
}
