% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbostat-pkg.R
\docType{package}
\name{npcharts-package}
\alias{npcharts}
\alias{npcharts-package}
\title{npcharts: NPC control charts}
\description{
The npcharts R package have been designed with the specific goal of making calibration and evaluation of multi-aspects NPC charts possible in practice.
}
