% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{npcharts.setseed}
\alias{npcharts.setseed}
\title{Set the seed used by this R packet}
\usage{
npcharts.setseed(seed)
}
\arguments{
\item{seed}{The seed. Either an integer scalar or an integer vector of length 2 representing a 64-bit seed.}
}
\description{
Set the seed used by this R packet
}
