% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calibrate.unconditional}
\alias{calibrate.unconditional}
\title{Unconditional calibration}
\usage{
calibrate.unconditional(
  m,
  n,
  distribution_key,
  distribution_parameters,
  monitoring_statistic_key,
  monitoring_statistic_parameters,
  limits_seq,
  is_upper_limit,
  nsim,
  run_length_cap
)
}
\arguments{
\item{m}{The dimension used for the reference sample.}

\item{n}{The dimension used for the test samples.}

\item{distribution_key}{A string which identifies a distribution or a distribution family. Used in combination with the "distribution_parameters" parameter in order to select the IC distribution.}

\item{distribution_parameters}{A list with the distribution parameters (an empty list should be supplied if the distribution has not parameters).}

\item{monitoring_statistic_key}{A string used to select the monitoring statistic of the chart to be calibrated.}

\item{monitoring_statistic_parameters}{A list with the parameters to be supplied for the selected monitoring statistic.}

\item{limits_seq}{The numeric vector of limits for which the run length should be recorded at each simulation.}

\item{is_upper_limit}{A boolean parameter used to select the direction of the OOC alternative.}

\item{nsim}{The number of simulations}

\item{run_length_cap}{A limit for the run length in the simulations used to guarantee convergence of the algorithm.}
}
\value{
A numeric matrix of size nsim x length(limits_seq).
}
\description{
Unconditional calibration for the Stehwart-type charts implemented by this package
}
