% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_monitoring_statistic}
\alias{compute_monitoring_statistic}
\title{Phase II monitoring}
\usage{
compute_monitoring_statistic(
  x1,
  x2,
  monitoring_statistic_type,
  monitoring_statistic_parameters
)
}
\arguments{
\item{monitoring_statistic_type}{Either "npc" or "simple_statistic"}

\item{monitoring_statistic_parameters}{Type-specific parameters of the monitoring statistic. For the "simple_statistic" type
the only parameter is "statistic" which should be set to a valid statistic key (see package doc). For the "npc" type the three
(required) parameters are "statistics", "permutation_distributions" and "tails".}

\item{m}{The dimension used for the reference sample.}

\item{n}{The dimension used for the test samples.}
}
\description{
Compute the observed value of the monitoring statistic
}
