% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_monitoring_statistic}
\alias{compute_monitoring_statistic}
\title{Phase II monitoring}
\usage{
compute_monitoring_statistic(
  x1,
  x2,
  monitoring_statistic_type,
  monitoring_statistic_parameters
)
}
\arguments{
\item{x1}{The reference sample}

\item{x2}{The test sample}

\item{monitoring_statistic_type}{Either "npc" or "simple_statistic"}

\item{monitoring_statistic_parameters}{A list with type-specific parameters of the monitoring statistic. For the
"simple_statistic" type the only parameter is "statistic" which should be set to a valid statistic key.
For the "npc" type the four (required) parameters are "statistics", "permutation_distributions", "tails" and
"combining_function". The "statistics", "permutation_distributions" and "tails" parameters should be three lists of the
same length. The "combining_function" parameters should be a string. See details for consulting the available statistics and combining
functions and the keys to be used.}
}
\value{
the monitoring statistic value
}
\description{
Compute the observed value of the monitoring statistic
}
\details{
The supported statistics and the respective keys to be used are:
\itemize{
 \item{Wilcoxon rank sum - "wilcoxon_rank_sum"}
 \item{Mann-Whitney - "mann_whitney"}
 \item{Mean-normalized wilcoxon rank sum  - "centered_wilcoxon_rank_sum"}
 \item{ Percentile modified rank test with s=r=0.42 - "percentiles_42_42"}
 \item{ Percentile modified rank test with s=r=0.46 - "percentiles_46_46"}
 \item{Klotz - "klotz_statistic"}
 \item{Ansari Bradely - "ab_statistic"}
 \item{Lepage - "lepage"}
 \item{Cucconi - "cucconi"}
}
The three options for the values of the "tails" list parameter of npc monitoring statistics are: "two_sided", "left" and "right"
The three options for the "combining_function" parameter of npc monitoring statistics are: "fisher", "liptak" and "tippet"
}
